Actualización 23-oct-2025

+++ routes/api.php
@@ -5,12 +5,14 @@ use App\Http\Controllers\API\ParteController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-Route::middleware('auth:sanctum')->group( function () {
-    Route::get('partes', [ParteController::class,'index']);
-    Route::post('parte', [ParteController::class,'store']);
-    Route::get('parte/{id}', [ParteController::class,'show']);
-    Route::put('parte/{id}', [ParteController::class,'update']);
-    Route::delete('parte/{id}', [ParteController::class,'destroy']);
+Route::middleware(['auth:sanctum','midread','mid1'])->group( function () {
+    Route::get('partes', [ParteController::class,'index'])->middleware('alguna:read,update');
+    Route::get('parte/{id}', [ParteController::class,'show'])
+        ->middleware(['midread','midmindundi']);
+        //->middleware('todas:read,admin');
+    Route::post('parte', [ParteController::class,'store'])->middleware(['midadmin']);
+    Route::put('parte/{id}', [ParteController::class,'update'])->middleware('midadmin');
+    Route::delete('parte/{id}', [ParteController::class,'destroy'])->middleware('midadmin');
 });
 
 Route::post('login', [AuthController::class, 'login']);
